// src/pages/Dashboard/Management/Attorney/components/Modals/PermissaoSigilo/interfaces/permissaoSigilo.interface.ts
export interface SubmitPermissaoSigilo {
  nuSigilo: { label: string; value: number } | null;
  idUsuario0: { label: string; value: number } | null;
  idUsuario1: { label: string; value: number } | null;
  idUsuario2: { label: string; value: number } | null;
  idUsuario3: { label: string; value: number } | null;
  idUsuario4: { label: string; value: number } | null;
  idUsuario5: { label: string; value: number } | null;
}